requirements.txt
asgiref==3.8.1
attrs==23.2.0
black==24.4.2
cfgv==3.4.0
click==8.1.7
coverage==7.5.4
distlib==0.3.8
Django==5.0.6
django-debug-toolbar==4.4.2
django-environ==0.11.2
django-filter==24.2
django-js-asset==2.2.0
django-mptt==0.16.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
factory-boy==3.3.0
Faker==26.0.0
filelock==3.15.4
flake8==7.1.0
identify==2.5.36
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
Markdown==3.6
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.7.1
pycodestyle==2.12.0
pyflakes==3.2.0
pytest==8.2.2
pytest-cov==5.0.0
pytest-django==4.8.0
pytest-factoryboy==2.7.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
uritemplate==4.1.1
virtualenv==20.26.3
